requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
scikit-image>=0.19.0
matplotlib>=3.4.0
scipy>=1.7.0
PyYAML>=5.4.0
Pillow>=8.3.0
tqdm>=4.62.0
